Refactored Target Event Handling to Eliminate Code Duplication in Action Classes
